rewrite frontend and Added several interfaces to the backend

+++ diskcloud/models/file.py
@@ -20,13 +20,13 @@ def download(path):
             Path(complete_path.parent).mkdir(parents=True,exist_ok=True)
             try:
                 with open(complete_path_str,'x') as tar:
-                    tar.add(Path(current_app.config['FILES_FOLDER'],'user',path).as_posix(), arcname = complete_path.name)
+                    tar.add(Path(current_app.config['FILES_FOLDER'],'user',path).as_posix(), arcname = complete_path.stem)
             except FileExistsError:
                 from os import remove
 
                 remove(complete_path_str)
                 with open(complete_path_str,'x') as tar:
-                    tar.add(Path(current_app.config['FILES_FOLDER'],'user',path).as_posix(), arcname = complete_path.name)
+                    tar.add(Path(current_app.config['FILES_FOLDER'],'user',path).as_posix(), arcname = complete_path.stem)
             if current_app.config['TESTING'] == False:
                 content_length = get_size(complete_path)
                 content_mime = "application/x-tar"
@@ -51,30 +51,28 @@ def download(path):
 
 def get_folder_content(username, path, name):
     from pathlib import Path
-    from flask import current_app
     from diskcloud.models.mysql import select_execute
 
     path = Path(path, name).as_posix()
     dir_list = []
     file_list = []
-    result = select_execute('select name, mtime from storage where username = %s and path = %s and type = %s', (username, path, 1))
+    result = select_execute('select type, name, modify_time, star, size from storage where username = %s and path = %s and trash_can = %s', (username, path, 0))
     for i in range(len(result)):
-        dir_list.append([result[i][0], result[i][1]])
-    result = select_execute('select name, mtime, size from storage where username = %s and path = %s and type = %s', (username, path, 0))
-    for i in range(len(result)):
-        file_list.append([result[i][0], result[i][1], result[i][2]])
+        if result[i][0] == 1:
+            dir_list.append([result[i][1], result[i][2], result[i][3]])
+        else:
+            file_list.append([result[i][1], result[i][2], result[i][3], result[i][4]])
     json_obj = {'directories': dir_list,'files': file_list}
     return json_obj
 
 def get_whole_folder_info(username):
-    from pathlib import Path
-    from flask import current_app
     from diskcloud.models.mysql import select_execute
+    from pathlib import Path
 
     def walk(username, path):
         dir_list = []
         dir_dict = {}
-        result = select_execute('select name from storage where username = %s and path = %s and type = %s', (username, path, 1))
+        result = select_execute('select name from storage where username = %s and path = %s and type = %s and trash_can = %s', (username, path, 1, 0))
         for i in range(len(result)):
             dir_list.append(result[i][0])
         if len(dir_list) != 0:
@@ -84,6 +82,63 @@ def get_whole_folder_info(username):
         return '0'
 
     json_obj = walk(username, '.')
+    return json_obj
+
+def get_star_info(username):
+    from diskcloud.models.mysql import select_execute
+
+    dir_list = []
+    file_list = []
+    result = select_execute('select type, name, star_time, size, path from storage where username = %s and star = %s and trash_can = %s', (username, 1, 0))
+    for i in range(len(result)):
+        if result[i][0] == 1:
+            dir_list.append([result[i][1], result[i][2], result[i][4]])
+        else:
+            file_list.append([result[i][1], result[i][2], result[i][3], result[i][4]])
+    json_obj = {'directories': dir_list, 'files': file_list}
+    return json_obj
+
+def get_share_info(username):
+    from diskcloud.models.mysql import select_execute
+
+    dir_list = []
+    file_list = []
+    result = select_execute('select type, name, share_time, expire_time, star, path from storage where username = %s and share = %s and trash_can = %s', (username, 1, 0))
+    for i in range(len(result)):
+        if result[i][0] == 1:
+            dir_list.append([result[i][1], result[i][2], result[i][3], result[i][4], result[i][5]])
+        else:
+            file_list.append([result[i][1], result[i][2], result[i][3], result[i][4], result[i][5]])
+    json_obj = {'directories': dir_list, 'files': file_list}
+    return json_obj
+
+def get_trash_can_info(username):
+    from diskcloud.models.mysql import select_execute
+
+    dir_list = []
+    file_list = []
+    result = select_execute('select type, name, trash_can_time, star, size from storage where username = %s and trash_can = %s', (username, 1))
+    for i in range(len(result)):
+        if result[i][0] == 1:
+            dir_list.append([result[i][1], result[i][2], result[i][3]])
+        else:
+            file_list.append([result[i][1], result[i][2], result[i][3], result[i][4]])
+    json_obj = {'directories': dir_list, 'files': file_list}
+    return json_obj
+
+def get_search_info(username, search_name):
+    from diskcloud.models.mysql import select_execute
+    from diskcloud.odels.string import name_parse
+
+    dir_list = []
+    file_list = []
+    result = select_execute('select type, name, modify_time, star, size, path from storage where username = %s and trash_can = %s and name like %s', (username, 0, name_parse(search_name)))
+    for i in range(len(result)):
+        if result[i][0] == 1:
+            dir_list.append([result[i][1], result[i][2], result[i][3], result[i][5]])
+        else:
+            file_list.append([result[i][1], result[i][2], result[i][3], result[i][4], result[i][5]])
+    json_obj = {'directories': dir_list, 'files': file_list}
     return json_obj
 
 def rename(username, path, name, af_name):
@@ -150,31 +205,39 @@ def delete(username, path, name, is_file):
             return True
         return False
 
-    if is_file is True:
-        if delete_db(username, path, name):
-            path = Path(current_app.config['FILES_FOLDER'], 'user', username, path, name).as_posix()
-            try:
-                remove(path)
-            except:
-                db_rollback()
-                return gen_error_res('fail to delete',500)
-            db_commit()
-            return ('', 200)
+    def in_trash_can(username, path, name):
+        result = select_execute('select trash_can from storage where username = %s and path = %s and name = %s', (username, path, name))
+        if result[0][0] == '1':
+            return True
+        return False
+
+    if in_trash_can(username, path, name):
+        if is_file is True:
+            if delete_db(username, path, name):
+                path = Path(current_app.config['FILES_FOLDER'], 'user', username, path, name).as_posix()
+                try:
+                    remove(path)
+                except:
+                    db_rollback()
+                    return gen_error_res('fail to delete',500)
+                db_commit()
+                return ('', 200)
+            else:
+                return gen_error_res('fail to update datebase', 500)
         else:
-            return gen_error_res('fail to update datebase', 500)
-    else:
-        if delete_folder_db(username, path, name):
-            path = Path(current_app.config['FILES_FOLDER'], 'user', username, path, name).as_posix()
-            try:
-                rmtree(path)
-            except:
+            if delete_folder_db(username, path, name):
+                path = Path(current_app.config['FILES_FOLDER'], 'user', username, path, name).as_posix()
+                try:
+                    rmtree(path)
+                except:
+                    db_rollback()
+                    return gen_error_res('fail to delete',500)
+                db_commit()
+                return ('', 200)
+            else:
                 db_rollback()
-                return gen_error_res('fail to delete',500)
-            db_commit()
-            return ('', 200)
-        else:
-            db_rollback()
-            return gen_error_res('fail to update datebase', 500)
+                return gen_error_res('fail to update datebase', 500)
+    return gen_error_res("can't delete,because this file/folder not in trash can.")
 
 def create_file(username, path, name, filename):
     from pathlib import Path
@@ -183,7 +246,7 @@ def create_file(username, path, name, filename):
     from diskcloud.models.time import now_time_str
     from diskcloud.models.mysql import insert_execute, db_commit, db_rollback
 
-    if insert_execute('insert into storage(username, path, name, size, mtime, type) values(%s, %s, %s, %s, %s, %s)', (username, Path(path, name).as_posix(), filename, 0, now_time_str(), 0)):
+    if insert_execute('insert into storage(username, path, name, size, modify_time, type) values(%s, %s, %s, %s, %s, %s)', (username, Path(path, name).as_posix(), filename, 0, now_time_str(), 0)):
         try:
             path = Path(current_app.config['FILES_FOLDER'], 'user', username, path, name, filename).as_posix()
             file = open(path,'a+')
@@ -205,7 +268,7 @@ def create_folder(username, path, name, foldername):
     from diskcloud.models.time import now_time_str
     from diskcloud.models.mysql import insert_execute, db_commit, db_rollback
 
-    if insert_execute('insert into storage(username, path, name, mtime, type) values(%s, %s, %s, %s, %s)', (username, Path(path, name).as_posix(), foldername, now_time_str(), 1)):
+    if insert_execute('insert into storage(username, path, name, modify_time, type) values(%s, %s, %s, %s, %s)', (username, Path(path, name).as_posix(), foldername, now_time_str(), 1)):
         path = Path(current_app.config['FILES_FOLDER'], 'user', username, path, name, foldername).as_posix()
         try:
             mkdir(path)
@@ -231,10 +294,56 @@ def save_file(username, path, name, file):
         file.save(whole_path_str)
     except:
         return gen_error_res('fail to save uploaded file', 500)
-    if insert_execute('insert into storage(username, path, name, size, mtime, type) values(%s, %s, %s, %s, %s, %s)', (username, Path(path, name).as_posix(), file.filename, whole_path.stat().st_size, now_time_str(), 0)):
+    if insert_execute('insert into storage(username, path, name, size, modify_time, type) values(%s, %s, %s, %s, %s, %s)', (username, Path(path, name).as_posix(), file.filename, whole_path.stat().st_size, now_time_str(), 0)):
         db_commit()
         return True
     return gen_error_res('fail to insert data to datebase', 500)
+
+def star(username, path, name):
+    from diskcloud.models.mysql import update_execute, db_commit, db_rollback
+    from diskcloud.models.response import gen_error_res
+    from diskcloud.models.time import now_time_str
+
+    result = update_execute('update storage set star = %s, star_time = %s where username = %s and path = %s and name = %s', (1, now_time_str(), username, path, name))
+    if result:
+        db_commit()
+        return ('', 200)
+    db_rollback()
+    return gen_error_res('fail to update datebase', 500)
+
+def unstar(username, path, name):
+    from diskcloud.models.mysql import update_execute, db_commit, db_rollback
+    from diskcloud.models.response import gen_error_res
+
+    result = update_execute('update storage set star = %s, star_time = %s where username = %s and path = %s and name = %s', (0, None, username, path, name))
+    if result:
+        db_commit()
+        return ('', 200)
+    db_rollback()
+    return gen_error_res('fail to update datebase', 500)
+
+def trash_can(username, path, name):
+    from diskcloud.models.mysql import update_execute, db_commit, db_rollback
+    from diskcloud.models.response import gen_error_res
+    from diskcloud.models.time import now_time_str
+
+    result = update_execute('update storage set trash_can = %s, trash_can_time = %s where username = %s and path = %s and name = %s', (1, now_time_str(), username, path, name))
+    if result:
+        db_commit()
+        return ('', 200)
+    db_rollback()
+    return gen_error_res('fail to update datebase', 500)
+
+def untrash_can(username, path, name):
+    from diskcloud.models.mysql import update_execute, db_commit, db_rollback
+    from diskcloud.models.response import gen_error_res
+
+    result = update_execute('update storage set trash_can = %s, trash_can_time = %s where username = %s and path = %s and name = %s', (0, None, username, path, name))
+    if result:
+        db_commit()
+        return ('', 200)
+    db_rollback()
+    return gen_error_res('fail to update datebase', 500)
 
 def get_mime(path):
     from .mime import MIME
